+++ crates/ruma-client-api/src/account/request_password_change_token_via_email.rs
@@ -10,7 +10,7 @@ pub mod v3 {
     use js_int::UInt;
     use ruma_common::{
         OwnedClientSecret, OwnedSessionId,
-        api::{auth_scheme::NoAuthentication, request, response},
+        api::{auth_scheme::NoAccessToken, request, response},
         metadata,
     };
 
@@ -19,7 +19,7 @@ pub mod v3 {
     metadata! {
         method: POST,
         rate_limited: false,
-        authentication: NoAuthentication,
+        authentication: NoAccessToken,
         history: {
             1.0 => "/_matrix/client/r0/account/password/email/requestToken",
             1.1 => "/_matrix/client/v3/account/password/email/requestToken",