+++ crates/ruma-client-api/src/discovery/discover_homeserver.rs
@@ -10,7 +10,7 @@ use std::borrow::Cow;
 #[cfg(feature = "unstable-msc4143")]
 use ruma_common::serde::JsonObject;
 use ruma_common::{
-    api::{auth_scheme::NoAuthentication, request, response},
+    api::{auth_scheme::NoAccessToken, request, response},
     metadata,
 };
 #[cfg(feature = "unstable-msc4143")]
@@ -22,7 +22,7 @@ use serde_json::Value as JsonValue;
 metadata! {
     method: GET,
     rate_limited: false,
-    authentication: NoAuthentication,
+    authentication: NoAccessToken,
     path: "/.well-known/matrix/client",
 }