+++ crates/ruma-client-api/src/media/get_content_thumbnail.rs
@@ -14,7 +14,7 @@ pub mod v3 {
     pub use ruma_common::media::Method;
     use ruma_common::{
         IdParseError, MxcUri, OwnedServerName,
-        api::{auth_scheme::NoAuthentication, request, response},
+        api::{auth_scheme::NoAccessToken, request, response},
         http_headers::ContentDisposition,
         metadata,
     };
@@ -24,7 +24,7 @@ pub mod v3 {
     metadata! {
         method: GET,
         rate_limited: true,
-        authentication: NoAuthentication,
+        authentication: NoAccessToken,
         history: {
             1.0 => "/_matrix/media/r0/thumbnail/{server_name}/{media_id}",
             1.1 => "/_matrix/media/v3/thumbnail/{server_name}/{media_id}",