+++ crates/ruma-client-api/src/uiaa/get_uiaa_fallback_page.rs
@@ -8,7 +8,7 @@ pub mod v3 {
     //! [spec]: https://spec.matrix.org/latest/client-server-api/#fallback
 
     use ruma_common::{
-        api::{auth_scheme::NoAuthentication, request},
+        api::{auth_scheme::NoAccessToken, request},
         metadata,
     };
 
@@ -17,7 +17,7 @@ pub mod v3 {
     metadata! {
         method: GET,
         rate_limited: false,
-        authentication: NoAuthentication,
+        authentication: NoAccessToken,
         history: {
             1.0 => "/_matrix/client/r0/auth/{auth_type}/fallback/web",
             1.1 => "/_matrix/client/v3/auth/{auth_type}/fallback/web",